Centralize app config init
