+++ crates/shared/src/config.rs
@@ -1,14 +1,29 @@
-//! Settings configuration.
+//! App configuration from *env*, *config* and *settings*.
+//!
+//! - *env* - Defined from CLI, env variables and "env-save.env" files, aggregated in [`crate::env::args`].
+//! - *config* - Defined in app resources and user config, not defined by the user directly (things like window state).
+//! - *settings* - Config the user can edit directly in the settings screen.
 
-use zng::config::*;
+use zng::{config::*, prelude::*};
 
-// others formats are available as Cargo features
-type FileConfig = JsonConfig;
-const CONFIG_FILE: &str = "config.json";
-const SETTINGS_FILE: &str = "settings.json";
+pub fn init() {
+    init_config();
+    init_l10n();
+    init_render();
+    #[cfg(feature = "release")]
+    init_licenses();
 
-/// Initialize user config in the app context.
-pub fn app_init() {
+    lang::bind();
+}
+
+/// Load config and settings.
+fn init_config() {
+    // others formats are available as Cargo features
+    type FileConfig = JsonConfig;
+
+    const CONFIG_FILE: &str = "config.json";
+    const SETTINGS_FILE: &str = "settings.json";
+
     // configs the user does not edit directly
     let default_config = FileConfig::read(zng::env::res(CONFIG_FILE));
     let user_config = FileConfig::sync(zng::env::config(CONFIG_FILE));
@@ -28,11 +43,43 @@ pub fn app_init() {
             .with_prefix("settings.", settings)
             .with_prefix("", config),
     );
-
-    lang::init();
 }
 
-/// Lang config
+/// Load localization.
+fn init_l10n() {
+    L10N.load_dir(&crate::env::args().lang_dir); // !!: TODO embedded setup
+}
+
+/// Configure render.
+fn init_render() {
+    // set render mode
+    WINDOWS
+        .default_render_mode()
+        .set(crate::env::args().render_mode);
+
+    // disable shader cache
+    if crate::env::args().no_shader_cache {
+        WINDOWS.default_cache_shaders().set(false);
+    }
+
+    // use ANGLE in all app windows
+    #[cfg(windows)]
+    if !crate::env::args().no_angle {
+        zng_view_angle::register_root_extender();
+    }
+}
+
+/// Load licenses.
+#[cfg(feature = "release")]
+fn init_licenses() {
+    // register bundled licenses, used by the default `OPEN_LICENSES_CMD` screen.
+    zng::third_party::LICENSES.register(crate::res::licenses);
+
+    #[cfg(windows)]
+    zng_view_angle::register_license();
+}
+
+/// Lang setting.
 pub mod lang {
     use zng::{l10n::Lang, prelude::*};
 
@@ -41,8 +88,8 @@ pub mod lang {
     /// Config placeholder for [`L10N::sys_lang`].
     pub const SYSTEM_LANG: Lang = lang!("system");
 
-    /// Bind the `L10n.app_lang` to the setting.
-    pub(crate) fn init() {
+    /// Bind `L10n.app_lang` to the setting.
+    pub(super) fn bind() {
         let actual_lang = expr_var! {
             let lang = #{CONFIG.get(CONFIG_KEY, SYSTEM_LANG)};
             if lang == &SYSTEM_LANG {
