+++ crates/shared/src/lib.rs
@@ -1,5 +1,3 @@
 pub mod config;
 pub mod env;
-pub mod l10n;
-pub mod render_cfg;
 pub mod res;
